update authors and add tests
