+++ install/script.py
@@ -473,5 +473,8 @@ env.addPackage('gEMpicker_v1.1',
                tar='gEMpicker_v1.1.tgz',
                default=False)
 
+env.addPackage('Gctf_v0.38',
+               tar='Gctf_v0.38.tgz',
+               default=False)
 
 env.execute()
